+++ iOS7/Frameworks/AVFoundation/AVAssetWriterClientInitiatedTerminalHelper.h
@@ -8,11 +8,11 @@
 
 @interface AVAssetWriterClientInitiatedTerminalHelper : AVAssetWriterTerminalHelper
 {
-    int _terminalStatus;
+    long long _terminalStatus;
 }
 
-- (int)status;
-- (id)initWithConfigurationState:(id)arg1 terminalStatus:(int)arg2;
+- (long long)status;
+- (id)initWithConfigurationState:(id)arg1 terminalStatus:(long long)arg2;
 - (id)initWithConfigurationState:(id)arg1;
 
 @end
