+++ iOS7/Frameworks/AVFoundation/AVPlayerItemAccessLog.h
@@ -17,7 +17,7 @@
 
 - (id)_accessLogArray;
 @property(readonly, nonatomic) NSArray *events;
-- (unsigned int)extendedLogDataStringEncoding;
+- (unsigned long long)extendedLogDataStringEncoding;
 - (id)extendedLogData;
 - (void)dealloc;
 - (void)finalize;
