+++ iOS7/Frameworks/AVFoundation/AVUnreachableAssetInspectorLoader.h
@@ -10,14 +10,14 @@
 
 @interface AVUnreachableAssetInspectorLoader : AVAssetInspectorLoader
 {
-    long _figError;
+    int _figError;
     NSDictionary *_figErrorUserInfo;
 }
 
 - (void)loadValuesAsynchronouslyForKeys:(id)arg1 completionHandler:(id)arg2;
-- (int)statusOfValueForKey:(id)arg1 error:(id *)arg2;
+- (long long)statusOfValueForKey:(id)arg1 error:(id *)arg2;
 - (void)dealloc;
-- (id)initWithFigError:(long)arg1 userInfo:(id)arg2;
+- (id)initWithFigError:(int)arg1 userInfo:(id)arg2;
 
 @end
