+++ iOS7/Frameworks/Accounts/ACDispatchCerberus.h
@@ -6,11 +6,12 @@
 
 #import "NSObject.h"
 
+// Not exported
 @interface ACDispatchCerberus : NSObject
 {
 }
 
-+ (void)waitOnSemaphore:(id)arg1 withTimeout:(unsigned int)arg2;
++ (void)waitOnSemaphore:(id)arg1 withTimeout:(unsigned long long)arg2;
 
 @end
