+++ iOS7/Frameworks/AddressBookUI/ABPrivacyAdditions.h
@@ -10,7 +10,7 @@
 {
 }
 
-+ (void)updateAccessDeniedView:(id)arg1 forAuthorizationStatus:(long)arg2;
++ (void)updateAccessDeniedView:(id)arg1 forAuthorizationStatus:(long long)arg2;
 
 @end
