+++ iOS7/Frameworks/AddressBookUI/ABPropertyFaceTimeAction.h
@@ -8,10 +8,10 @@
 
 @interface ABPropertyFaceTimeAction : ABPropertyAction
 {
-    int _type;
+    long long _type;
 }
 
-@property(nonatomic) int type; // @synthesize type=_type;
+@property(nonatomic) long long type; // @synthesize type=_type;
 - (void)performActionForItem:(id)arg1 sender:(id)arg2;
 
 @end
