+++ iOS7/Frameworks/AddressBookUI/ABRepeatingGradientSeparatorView.h
@@ -13,10 +13,10 @@
     NSArray *_gradientColors;
     UIColor *_startColor;
     UIColor *_endColor;
-    float _step;
+    double _step;
 }
 
-@property(nonatomic) float step; // @synthesize step=_step;
+@property(nonatomic) double step; // @synthesize step=_step;
 @property(retain, nonatomic) UIColor *endColor; // @synthesize endColor=_endColor;
 @property(retain, nonatomic) UIColor *startColor; // @synthesize startColor=_startColor;
 @property(readonly, nonatomic) NSArray *gradientColors;
