+++ iOS7/Frameworks/AddressBookUI/UIModalViewDelegate-Protocol.h
@@ -9,11 +9,11 @@
 @protocol UIModalViewDelegate <NSObject>
 
 @optional
-- (void)modalView:(id)arg1 didDismissWithButtonIndex:(int)arg2;
-- (void)modalView:(id)arg1 willDismissWithButtonIndex:(int)arg2;
+- (void)modalView:(id)arg1 didDismissWithButtonIndex:(long long)arg2;
+- (void)modalView:(id)arg1 willDismissWithButtonIndex:(long long)arg2;
 - (void)didPresentModalView:(id)arg1;
 - (void)willPresentModalView:(id)arg1;
 - (void)modalViewCancel:(id)arg1;
-- (void)modalView:(id)arg1 clickedButtonAtIndex:(int)arg2;
+- (void)modalView:(id)arg1 clickedButtonAtIndex:(long long)arg2;
 @end
