+++ iOS7/Frameworks/CoreData/NSSQLStoreMappingGenerator.h
@@ -8,6 +8,7 @@
 
 @class NSMutableDictionary;
 
+// Not exported
 @interface NSSQLStoreMappingGenerator : NSObject
 {
     NSMutableDictionary *_names;
@@ -17,7 +18,7 @@
 + (id)defaultMappingGenerator;
 - (id)generateTableName:(id)arg1;
 - (id)newGeneratedPropertyName:(id)arg1;
-- (id)newUniqueNameWithBase:(unsigned short *)arg1 withLength:(unsigned int)arg2;
+- (id)newUniqueNameWithBase:(unsigned short *)arg1 withLength:(unsigned long long)arg2;
 - (id)uniqueNameWithBase:(id)arg1;
 - (void)dealloc;
 - (id)init;
