+++ iOS7/Frameworks/CoreData/PFUbiquityMergeConflict.h
@@ -8,6 +8,7 @@
 
 @class NSDictionary;
 
+// Not exported
 @interface PFUbiquityMergeConflict : NSMergeConflict
 {
     NSDictionary *_ancestorSnapshot;
@@ -15,7 +16,7 @@
 
 @property(retain) NSDictionary *ancestorSnapshot; // @synthesize ancestorSnapshot=_ancestorSnapshot;
 - (void)dealloc;
-- (id)initWithSource:(id)arg1 newVersion:(unsigned int)arg2 oldVersion:(unsigned int)arg3 cachedSnapshot:(id)arg4 persistedSnapshot:(id)arg5;
+- (id)initWithSource:(id)arg1 newVersion:(unsigned long long)arg2 oldVersion:(unsigned long long)arg3 cachedSnapshot:(id)arg4 persistedSnapshot:(id)arg5;
 
 @end
