+++ iOS7/Frameworks/CoreData/PFZipEndOfCentralDirectoryRecord.h
@@ -8,6 +8,7 @@
 
 @class NSString;
 
+// Not exported
 @interface PFZipEndOfCentralDirectoryRecord : NSObject
 {
     unsigned short _numberOfDisk;
@@ -29,8 +30,8 @@
 @property(nonatomic) unsigned short diskWhereCentralDirectoryStarts; // @synthesize diskWhereCentralDirectoryStarts=_diskWhereCentralDirectoryStarts;
 @property(nonatomic) unsigned short numberOfDisk; // @synthesize numberOfDisk=_numberOfDisk;
 - (void)appendToData:(id)arg1;
-- (unsigned int)loadFromBytes:(const char *)arg1 offset:(unsigned int)arg2;
-- (unsigned int)loadFromData:(id)arg1 offset:(unsigned int)arg2;
+- (unsigned long long)loadFromBytes:(const char *)arg1 offset:(unsigned long long)arg2;
+- (unsigned long long)loadFromData:(id)arg1 offset:(unsigned long long)arg2;
 - (id)description;
 - (void)dealloc;
 - (id)init;
