+++ iOS7/Frameworks/CoreData/_NSCoreDataException.h
@@ -8,19 +8,20 @@
 
 @class NSString;
 
+// Not exported
 @interface _NSCoreDataException : NSException
 {
-    long _code;
+    long long _code;
     NSString *_domain;
 }
 
-+ (id)exceptionWithName:(id)arg1 code:(long)arg2 reason:(id)arg3 userInfo:(id)arg4;
++ (id)exceptionWithName:(id)arg1 code:(long long)arg2 reason:(id)arg3 userInfo:(id)arg4;
 - (id)domain;
-- (long)code;
+- (long long)code;
 - (void)_setDomain:(id)arg1;
 - (void)dealloc;
 - (id)initWithName:(id)arg1 reason:(id)arg2 userInfo:(id)arg3;
-- (id)initWithName:(id)arg1 code:(long)arg2 reason:(id)arg3 userInfo:(id)arg4;
+- (id)initWithName:(id)arg1 code:(long long)arg2 reason:(id)arg3 userInfo:(id)arg4;
 
 @end
