+++ iOS7/Frameworks/CoreData/_NSManagedObject_st.h
@@ -8,18 +8,19 @@
 
 @class NSManagedObjectContext, NSManagedObjectID;
 
+// Not exported
 @interface _NSManagedObject_st : NSObject
 {
     int _cd_rc;
-    unsigned long _cd_stateFlags;
+    unsigned long long _cd_stateFlags;
     id _cd_rawData;
     id _cd_entity;
     NSManagedObjectContext *_cd_managedObjectContext;
     NSManagedObjectID *_cd_objectID;
-    unsigned long _cd_extraFlags;
+    unsigned long long _cd_extraFlags;
     id _cd_observationInfo;
     id *_cd_snapshots;
-    unsigned long _cd_lockingInfo;
+    unsigned long long _cd_lockingInfo;
     id _cd_queueReference;
 }
