+++ iOS7/Frameworks/CoreData/_NSOrderedSetDiffChange.h
@@ -6,12 +6,13 @@
 
 #import <CoreData/_NSOrderedSetDiff.h>
 
+// Not exported
 @interface _NSOrderedSetDiffChange : _NSOrderedSetDiff
 {
-    unsigned int _index;
+    unsigned long long _index;
 }
 
-- (id)initWithObject:(id)arg1 index:(unsigned int)arg2;
+- (id)initWithObject:(id)arg1 index:(unsigned long long)arg2;
 
 @end
