+++ iOS7/Frameworks/CoreFoundation/__NSGenericDeallocHandler.h
@@ -4,6 +4,7 @@
  *     class-dump is Copyright (C) 1997-1998, 2000-2001, 2004-2011 by Steve Nygard.
  */
 
+// Not exported
 @interface __NSGenericDeallocHandler
 {
     Class isa;
@@ -13,7 +14,7 @@
 + (void)initialize;
 - (void)finalize;
 - (void)release;
-- (unsigned int)retainCount;
+- (unsigned long long)retainCount;
 - (id)retain;
 
 @end
