+++ iOS7/Frameworks/CoreFoundation/__NSPlaceholderArray.h
@@ -6,6 +6,7 @@
 
 #import <CoreFoundation/NSMutableArray.h>
 
+// Not exported
 @interface __NSPlaceholderArray : NSMutableArray
 {
 }
@@ -14,20 +15,20 @@
 + (id)mutablePlaceholder;
 + (id)immutablePlaceholder;
 + (void)initialize;
-- (void)replaceObjectAtIndex:(unsigned int)arg1 withObject:(id)arg2;
-- (void)removeObjectAtIndex:(unsigned int)arg1;
-- (void)insertObject:(id)arg1 atIndex:(unsigned int)arg2;
-- (id)objectAtIndex:(unsigned int)arg1;
-- (unsigned int)count;
+- (void)replaceObjectAtIndex:(unsigned long long)arg1 withObject:(id)arg2;
+- (void)removeObjectAtIndex:(unsigned long long)arg1;
+- (void)insertObject:(id)arg1 atIndex:(unsigned long long)arg2;
+- (id)objectAtIndex:(unsigned long long)arg1;
+- (unsigned long long)count;
 - (void)dealloc;
-- (unsigned int)retainCount;
+- (unsigned long long)retainCount;
 - (oneway void)release;
 - (id)retain;
 - (id)initWithContentsOfURL:(id)arg1;
 - (id)initWithContentsOfFile:(id)arg1;
 - (id)init;
-- (id)initWithCapacity:(unsigned int)arg1;
-- (id)initWithObjects:(const id *)arg1 count:(unsigned int)arg2;
+- (id)initWithCapacity:(unsigned long long)arg1;
+- (id)initWithObjects:(const id *)arg1 count:(unsigned long long)arg2;
 
 @end
